+++ components/molecules/FeedbackWidget.js
@@ -201,22 +201,11 @@ export const FeedbackWidget = ({
                 <h2 className="text-h4 lg:text-h3 lg:text-sm font-display pt-6 mb-4 w-48 sm:w-auto">
                   {t("improveService")}
                 </h2>
-                <ul className="list-outside list-disc px-6 py-2">
-                  <li className="text-xs lg:text-sm font-body mb-4">
-                    <strong>{t("reportAProblemNoReply")}</strong>{" "}
-                    {t("reportAProblemEnquiries")}{" "}
-                    <ActionButton
-                      id="link-mail"
-                      ariaLabel="Service Canada email"
-                      dataCy="link-mail"
-                      dataTestId="link-mail"
-                      href={`mailto:${email}`}
-                      text={email}
-                      custom="text-xs lg:text-sm underline outline-none focus:outline-white-solid"
-                    />
-                    .
+                <ul className="list-outside list-disc px-6 pb-3">
+                  <li className="text-xs lg:text-sm pt-2 pb-1 font-body">
+                    <strong>{t("reportAProblemNoReply")}</strong>
                   </li>
-                  <li className="text-xs lg:text-sm font-body mb-4">
+                  <li className="text-xs lg:text-sm font-body mb-0">
                     <strong>{t("confidential")}</strong>
                     <ActionButton
                       ariaLabel="Privacy page link"